SAM files

The SAM.Record

The XAM package supports the following accessors for SAM.Record types.

XAM.SAM.Record — Type

SAM.Record()

Create an unfilled SAM record.

SAM.Record(data::Vector{UInt8})

Create a SAM record from data. This function verifies the format and indexes fields for accessors. Note that the ownership of data is transferred to a new record object.

SAM.Record(str::AbstractString)

Create a SAM record from str. This function verifies the format and indexes fields for accessors.

SAM.Reader and SAM.Writer

Reference:

XAM.SAM.Reader — Type

SAM.Reader(input::IO)

Create a data reader of the SAM file format.

Arguments

input: data source

Missing docstring.

Missing docstring for SAM.Writer. Check Documenter's build log for details.